+++ Mu3Library_Base/Runtime/Scripts/DI/IDICore.cs
@@ -1,0 +1,10 @@
+namespace Mu3Library.DI
+{
+    public interface IDICore
+    {
+        public bool IsContainerConfigured { get; }
+        public bool IsPrepared { get; }
+
+        public int ExecutionOrder { get; }
+    }
+}