+++ Mu3Library_Base/Runtime/Scripts/Scene/ISceneLoaderEventBus.cs
@@ -1,4 +1,5 @@
 using System;
+using Mu3Library.Foundation.Event;
 
 namespace Mu3Library.Scene
 {
@@ -93,26 +94,26 @@ namespace Mu3Library.Scene
 
         public event Action<SceneCommandRejectedInfo> OnSceneCommandRejected;
 
-        public uint SubscribeOnSingleSceneLoadStartedOnce(Action<string> callback);
-        public uint SubscribeOnSingleSceneLoadStartedOnce(Action<string> callback, Action onDisposed);
-        public uint SubscribeOnSingleScenePreloadedOnce(Action<string> callback);
-        public uint SubscribeOnSingleScenePreloadedOnce(Action<string> callback, Action onDisposed);
-        public uint SubscribeOnSingleSceneLoadedOnce(Action<string> callback);
-        public uint SubscribeOnSingleSceneLoadedOnce(Action<string> callback, Action onDisposed);
-        public uint SubscribeOnSingleSceneChangedOnce(Action<string, string> callback);
-        public uint SubscribeOnSingleSceneChangedOnce(Action<string, string> callback, Action onDisposed);
-        public uint SubscribeOnSingleSceneLifecycleOnce(Action<SceneLifecycleInfo> callback);
-        public uint SubscribeOnSingleSceneLifecycleOnce(Action<SceneLifecycleInfo> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnSingleSceneLoadStartedOnce(Action<string> callback);
+        public ISubscriptionInfo SubscribeOnSingleSceneLoadStartedOnce(Action<string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnSingleScenePreloadedOnce(Action<string> callback);
+        public ISubscriptionInfo SubscribeOnSingleScenePreloadedOnce(Action<string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnSingleSceneLoadedOnce(Action<string> callback);
+        public ISubscriptionInfo SubscribeOnSingleSceneLoadedOnce(Action<string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnSingleSceneChangedOnce(Action<string, string> callback);
+        public ISubscriptionInfo SubscribeOnSingleSceneChangedOnce(Action<string, string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnSingleSceneLifecycleOnce(Action<SceneLifecycleInfo> callback);
+        public ISubscriptionInfo SubscribeOnSingleSceneLifecycleOnce(Action<SceneLifecycleInfo> callback, Action onDisposed);
 
-        public uint SubscribeOnAdditiveSceneLoadStartedOnce(Action<string> callback);
-        public uint SubscribeOnAdditiveSceneLoadStartedOnce(Action<string> callback, Action onDisposed);
-        public uint SubscribeOnAdditiveScenePreloadedOnce(Action<string> callback);
-        public uint SubscribeOnAdditiveScenePreloadedOnce(Action<string> callback, Action onDisposed);
-        public uint SubscribeOnAdditiveSceneLoadedOnce(Action<string> callback);
-        public uint SubscribeOnAdditiveSceneLoadedOnce(Action<string> callback, Action onDisposed);
-        public uint SubscribeOnAdditiveSceneUnloadedOnce(Action<string> callback);
-        public uint SubscribeOnAdditiveSceneUnloadedOnce(Action<string> callback, Action onDisposed);
-        public uint SubscribeOnAdditiveSceneLifecycleOnce(Action<SceneLifecycleInfo> callback);
-        public uint SubscribeOnAdditiveSceneLifecycleOnce(Action<SceneLifecycleInfo> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneLoadStartedOnce(Action<string> callback);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneLoadStartedOnce(Action<string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnAdditiveScenePreloadedOnce(Action<string> callback);
+        public ISubscriptionInfo SubscribeOnAdditiveScenePreloadedOnce(Action<string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneLoadedOnce(Action<string> callback);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneLoadedOnce(Action<string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneUnloadedOnce(Action<string> callback);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneUnloadedOnce(Action<string> callback, Action onDisposed);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneLifecycleOnce(Action<SceneLifecycleInfo> callback);
+        public ISubscriptionInfo SubscribeOnAdditiveSceneLifecycleOnce(Action<SceneLifecycleInfo> callback, Action onDisposed);
     }
 }